refactor: 아이콘 파일을 SVG에서 PNG로 변경
